Checking in Manifold Cfg. New Environment Definition, and reduced Cfg classes. Please refer to email for more details.
git-svn-id: svn+ssh://parasol-svn.cs.tamu.edu/research/parasol-svn/svnrepository/pmpl/trunk@3267 5e2c2758-810a-0410-8eb3-8e4e92d4004e

+++ src/DistanceMetric/DistanceMetricMethod.cpp
@@ -30,14 +30,14 @@ void DistanceMetricMethod::ScaleCfg(Environment* _env, double _length, Cfg& _o, 
   Cfg* outsideCfg = _c.CreateNewCfg();
   // first find an outsite configuration with sufficient size
   while(Distance(_env, *origin, *outsideCfg) < 2*_length)
-    for(int i=0; i<outsideCfg->DOF(); ++i)
+    for(size_t i=0; i<outsideCfg->DOF(); ++i)
       outsideCfg->SetSingleParam(i, 2*outsideCfg->GetSingleParam(i), _norm);
   // now, using binary search  find a configuration with the approximate length
   Cfg* aboveCfg = outsideCfg->CreateNewCfg();
   Cfg* belowCfg = origin->CreateNewCfg();
   Cfg* currentCfg = _c.CreateNewCfg();
   while (1) {
-    for(int i=0; i<currentCfg->DOF(); ++i)
+    for(size_t i=0; i<currentCfg->DOF(); ++i)
       currentCfg->SetSingleParam(i, (aboveCfg->GetSingleParam(i) + belowCfg->GetSingleParam(i)) / 2.0, _norm);
     double magnitude = Distance(_env, *origin, *currentCfg);
     if( (magnitude >= _length*0.9) && (magnitude <= _length*1.1)) 
@@ -47,8 +47,10 @@ void DistanceMetricMethod::ScaleCfg(Environment* _env, double _length, Cfg& _o, 
     else 
       *belowCfg = *currentCfg; 
   }
-  for(int i=0; i<_c.DOF(); ++i)
+  
+  for(size_t i=0; i<_c.DOF(); ++i)
     _c.SetSingleParam(i, currentCfg->GetSingleParam(i), _norm);
+
   //delete origin;
   delete outsideCfg;
   delete aboveCfg;
