+++ src/DistanceMetric/EuclideanDistance.cpp
@@ -29,12 +29,12 @@ EuclideanDistance::ScaledDistanceImpl(Environment* _env,const Cfg& _c1, const Cf
   tmp.subtract(_c1,_c2);
   double posMag(0.0);
   double maxRange(0.0);
-  for(int i=0; i< tmp.PosDOF(); ++i) {
+  for(size_t i=0; i< tmp.PosDOF(); ++i) {
     std::pair<double,double> range = _env->GetBoundingBox()->GetRange(i);
     double tmpRange = range.second-range.first;
     if(tmpRange > maxRange) maxRange = tmpRange;
   }
-  for(int i=0; i< tmp.PosDOF(); ++i) {
+  for(size_t i=0; i< tmp.PosDOF(); ++i) {
      posMag += sqr(tmp.GetSingleParam(i) / maxRange);
   }
   double dReturn = sqrt(_sValue*posMag + (1.0 - _sValue)*sqr(tmp.OrientationMagnitude()) );
@@ -45,7 +45,7 @@ void EuclideanDistance::ScaleCfg(Environment* _env, double _length, Cfg& _o, Cfg
   double originalLength = this->Distance(_env, _o, _c);
   double diff;
   do {
-    for(int i=0; i<_c.DOF(); ++i)
+    for(size_t i=0; i<_c.DOF(); ++i)
       _c.SetSingleParam(i, (_length/originalLength)*_c.GetSingleParam(i), _norm);
     originalLength = this->Distance(_env, _o, _c);
     diff = _length - originalLength;
