+++ src/DistanceMetric/PureEuclideanDistance.cpp
@@ -14,7 +14,7 @@ void PureEuclideanDistance::PrintOptions(ostream& _os) const {
 
 double PureEuclideanDistance::Distance(Environment* _env, const Cfg& _c1, const Cfg& _c2) {
   double dist(0.0);
-  for(int i=0; i<_c1.DOF(); ++i) {
+  for(size_t i=0; i<_c1.DOF(); ++i) {
       double diff = _c1.GetSingleParam(i) - _c2.GetSingleParam(i);
       dist += diff * diff;
   }
